# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Username/password authentication operations
  - name: monitoring
    description: Monitoring toggle operations
  - name: trackedUsers
    description: Tracked user list operations
  - name: settings
    description: Outbound request configuration (target URL, headers, cookies)
  - name: automationJobs
    description: >-
      Stored automation job configuration. Jobs are persisted for record-keeping
      only - nothing in this codebase schedules or executes them. Any real
      execution against a third-party site must be triggered manually by a
      user, the same way /settings/request-config/test works.
paths:
  /settings/request-config:
    get:
      operationId: getRequestConfig
      tags: [settings]
      summary: Get the saved outbound request configuration
      responses:
        "200":
          description: Request configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestConfig"
    put:
      operationId: updateRequestConfig
      tags: [settings]
      summary: Save the outbound request configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestConfigInput"
      responses:
        "200":
          description: Updated request configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestConfig"
  /settings/request-config/test:
    post:
      operationId: testRequestConfig
      tags: [settings]
      summary: Send a real request to the saved target URL using the saved headers/cookies and return the response
      responses:
        "200":
          description: Result of the test request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestConfigTestResult"
        "400":
          description: No target URL configured, or the request failed before a response was received
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with a username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid username or password
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Log out the current session. Idempotent - returns 204 even if not authenticated.
      responses:
        "204":
          description: Logged out (or already logged out)
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the currently authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /automation-jobs:
    get:
      operationId: listAutomationJobs
      tags: [automationJobs]
      summary: List stored automation job configurations
      description: >-
        Read-only listing of persisted job configuration. No job in this list
        is ever executed automatically by the server.
      responses:
        "200":
          description: List of automation jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutomationJob"
    post:
      operationId: createAutomationJob
      tags: [automationJobs]
      summary: Save an automation job configuration
      description: >-
        Persists the job configuration only. Jobs are created with status
        "paused" and are never picked up by a scheduler - there is none in
        this codebase. A user must build/trigger any real execution
        explicitly elsewhere.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutomationJobInput"
      responses:
        "201":
          description: Created automation job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationJob"
  /automation-jobs/{jobId}:
    delete:
      operationId: deleteAutomationJob
      tags: [automationJobs]
      summary: Remove a stored automation job configuration
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /monitoring:
    get:
      operationId: getMonitoringStatus
      tags: [monitoring]
      summary: Get monitoring status
      responses:
        "200":
          description: Monitoring status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitoringStatus"
    patch:
      operationId: updateMonitoringStatus
      tags: [monitoring]
      summary: Enable or disable monitoring
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonitoringUpdate"
      responses:
        "200":
          description: Updated monitoring status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitoringStatus"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [trackedUsers]
      summary: Get counts per category
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /tracked-users:
    get:
      operationId: listTrackedUsers
      tags: [trackedUsers]
      summary: List tracked users
      description: Returns tracked users, optionally filtered by category
      parameters:
        - name: category
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TrackedUserCategory"
      responses:
        "200":
          description: List of tracked users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackedUser"
    post:
      operationId: createTrackedUser
      tags: [trackedUsers]
      summary: Add a tracked user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackedUserInput"
      responses:
        "201":
          description: Created tracked user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackedUser"
  /tracked-users/{id}:
    delete:
      operationId: deleteTrackedUser
      tags: [trackedUsers]
      summary: Remove a tracked user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
        lastRestart:
          type: string
          format: date-time
          description: When the current server process started.
      required:
        - status
        - lastRestart
    ActionType:
      type: string
      enum: [like, view_story, follow]
    AutomationJobInput:
      type: object
      required: [targetUsername, actionType, frequencyMinutes]
      properties:
        targetUsername:
          type: string
          minLength: 1
        actionType:
          $ref: "#/components/schemas/ActionType"
        frequencyMinutes:
          type: integer
          minimum: 15
          maximum: 1440
        randomizeDelay:
          type: boolean
          default: true
    AutomationJob:
      type: object
      required: [jobId, targetUsername, actionType, frequencyMinutes, randomizeDelay, status, nextRunAt, createdAt]
      properties:
        jobId:
          type: string
        targetUsername:
          type: string
        actionType:
          $ref: "#/components/schemas/ActionType"
        frequencyMinutes:
          type: integer
        randomizeDelay:
          type: boolean
        status:
          type: string
          enum: [active, paused, failed]
          description: Always created as "paused". Nothing in this codebase transitions it or runs the job.
        nextRunAt:
          type: string
          format: date-time
          description: Informational only (createdAt + frequencyMinutes) - not tied to any real scheduler.
        createdAt:
          type: string
          format: date-time
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
        deviceProfile:
          type: string
          description: Optional client-supplied device/browser identifier. Echoed back on login, not persisted or used server-side.
    AuthUser:
      type: object
      required: [id, username, sessionExpiry]
      properties:
        id:
          type: integer
        username:
          type: string
        sessionExpiry:
          type: string
          format: date-time
          description: When the current session cookie expires.
        deviceProfile:
          type: string
          nullable: true
          description: Optional client-supplied device/browser identifier, echoed back as received. Not used for anything server-side.
    MonitoringStatus:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
    MonitoringUpdate:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
    TrackedUserCategory:
      type: string
      enum: [follower, liked_post, liked_story]
    TrackedUser:
      type: object
      required: [id, username, fullName, avatarUrl, category, addedAt, interactionCount, autoLikeEnabled]
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        avatarUrl:
          type: string
        category:
          $ref: "#/components/schemas/TrackedUserCategory"
        addedAt:
          type: string
          format: date-time
        lastInteractionAt:
          type: string
          format: date-time
          nullable: true
          description: Set manually via the API - nothing in this codebase writes to it automatically.
        interactionCount:
          type: integer
          description: Set manually via the API - nothing in this codebase increments it automatically.
        autoLikeEnabled:
          type: boolean
          description: Stored preference flag only - no automated liking exists in this codebase.
    TrackedUserInput:
      type: object
      required: [username, fullName, avatarUrl, category]
      properties:
        username:
          type: string
          minLength: 1
        fullName:
          type: string
          minLength: 1
        avatarUrl:
          type: string
        category:
          $ref: "#/components/schemas/TrackedUserCategory"
        autoLikeEnabled:
          type: boolean
          default: false
          description: Stored preference flag only - no automated liking exists in this codebase.
    DashboardSummary:
      type: object
      required: [followerCount, likedPostCount, likedStoryCount, monitoringEnabled, rateLimitStatus]
      properties:
        followerCount:
          type: integer
        likedPostCount:
          type: integer
        likedStoryCount:
          type: integer
        monitoringEnabled:
          type: boolean
        rateLimitStatus:
          type: string
          enum: [safe, warning, critical]
          description: Static placeholder ("safe") - no real request-rate tracking exists in this codebase.
    RequestConfig:
      type: object
      required: [targetUrl, headers, cookies]
      properties:
        targetUrl:
          type: string
          nullable: true
        headers:
          type: object
          additionalProperties:
            type: string
        cookies:
          type: object
          additionalProperties:
            type: string
    RequestConfigInput:
      type: object
      required: [targetUrl, headers, cookies]
      properties:
        targetUrl:
          type: string
          nullable: true
        headers:
          type: object
          additionalProperties:
            type: string
        cookies:
          type: object
          additionalProperties:
            type: string
    RequestConfigTestResult:
      type: object
      required: [status, statusText, headers, bodyPreview]
      properties:
        status:
          type: integer
        statusText:
          type: string
        headers:
          type: object
          additionalProperties:
            type: string
        bodyPreview:
          type: string