# lib/api-zod/src/generated/types/actionType.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type ActionType = typeof ActionType[keyof typeof ActionType];


export const ActionType = {
  like: 'like',
  view_story: 'view_story',
  follow: 'follow',
} as const;