# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  id: number;
  username: string;
  /** When the current session cookie expires. */
  sessionExpiry: Date;
  /** Optional client-supplied device/browser identifier, echoed back as received. Not used for anything server-side. */
  deviceProfile?: string | null;
}