# lib/api-zod/src/generated/types/automationJob.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActionType } from './actionType';
import type { AutomationJobStatus } from './automationJobStatus';

export interface AutomationJob {
  jobId: string;
  targetUsername: string;
  actionType: ActionType;
  frequencyMinutes: number;
  randomizeDelay: boolean;
  /** Always created as "paused". Nothing in this codebase transitions it or runs the job. */
  status: AutomationJobStatus;
  /** Informational only (createdAt + frequencyMinutes) - not tied to any real scheduler. */
  nextRunAt: Date;
  createdAt: Date;
}