# lib/api-zod/src/generated/types/automationJobInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActionType } from './actionType';

export interface AutomationJobInput {
  /** @minLength 1 */
  targetUsername: string;
  actionType: ActionType;
  /**
     * @minimum 15
     * @maximum 1440
     */
  frequencyMinutes: number;
  randomizeDelay?: boolean;
}