# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryRateLimitStatus } from './dashboardSummaryRateLimitStatus';

export interface DashboardSummary {
  followerCount: number;
  likedPostCount: number;
  likedStoryCount: number;
  monitoringEnabled: boolean;
  /** Static placeholder ("safe") - no real request-rate tracking exists in this codebase. */
  rateLimitStatus: DashboardSummaryRateLimitStatus;
}