# lib/api-zod/src/generated/types/loginRequest.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LoginRequest {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
  /** Optional client-supplied device/browser identifier. Echoed back on login, not persisted or used server-side. */
  deviceProfile?: string;
}