# lib/api-zod/src/generated/types/trackedUser.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TrackedUserCategory } from './trackedUserCategory';

export interface TrackedUser {
  id: number;
  username: string;
  fullName: string;
  avatarUrl: string;
  category: TrackedUserCategory;
  addedAt: Date;
  /** Set manually via the API - nothing in this codebase writes to it automatically. */
  lastInteractionAt?: Date | null;
  /** Set manually via the API - nothing in this codebase increments it automatically. */
  interactionCount: number;
  /** Stored preference flag only - no automated liking exists in this codebase. */
  autoLikeEnabled: boolean;
}