# lib/api-zod/src/generated/types/trackedUserInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TrackedUserCategory } from './trackedUserCategory';

export interface TrackedUserInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  fullName: string;
  avatarUrl: string;
  category: TrackedUserCategory;
  /** Stored preference flag only - no automated liking exists in this codebase. */
  autoLikeEnabled?: boolean;
}